lesson7 is ready

+++ resources/js/app.js
@@ -1,4 +1,5 @@
 require("./bootstrap");
+import { result } from "lodash";
 import http from "./http";
 
 const adminWrapperNews = document.querySelector(".admin-wrapper-news");
@@ -7,8 +8,12 @@ adminWrapperNews?.addEventListener("click", async (e) => {
     const target = e.target;
 
     if (target.dataset.deleteId) {
-        await http.delete(
-            `/delete-new/${target.dataset.deleteId}`
-        );
+        const result = confirm("Are you sure?");
+        if (result) {
+            const response = await http.delete(
+                `/delete-new/${target.dataset.deleteId}`
+            );
+            window.location.reload();
+        }
     }
 });
